requirements.txt

Flask==2.2.5
pyotp==2.8.0
smartapi-python==1.5.5
